requirements.txt
glob2
numpy
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.13.1+cu113
--extra-index-url https://download.pytorch.org/whl/cu113
torchaudio==0.12.1 
gymnasium
wandb
swig
pygame
box2d
moviepy
pyyaml
matplotlib